And sort out the archivist fixtures
